% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_get_data.R
\name{cl_get_data}
\alias{cl_get_data}
\title{Retrieve UCDP Dataset from API}
\usage{
cl_get_data(
  resource,
  version,
  country = NULL,
  conflict = NULL,
  year = NULL,
  conflict_incompatibility = NULL,
  dyad = NULL,
  org = NULL,
  conflict_type = NULL,
  geography = NULL,
  start_date = NULL,
  end_date = NULL,
  type_of_violence = NULL,
  actor = NULL
)
}
\arguments{
\item{resource}{A character string specifying the type of resource to request.
Must be one of:
\itemize{
\item \code{"ucdpprioconflict"}: State-based conflicts data.
\item \code{"dyadic"}: Dyadic-level conflict data.
\item \code{"battledeaths"}: Data on battle-related deaths.
\item \code{"nonstate"}: Non-state conflicts data.
\item \code{"onesided"}: Data on one-sided violence.
\item \code{"gedevents"}: Event-based conflict data from the UCDP GED.
}}

\item{version}{A character string specifying the version of the resource to request.}

\item{country}{A vector of numeric country codes (Gleditsch and Ward codes).
\itemize{
\item Valid for all resources.
\item Must be a numeric vector, where each code corresponds to a country involved in the conflict.
}}

\item{conflict}{A vector of numeric conflict IDs.
\itemize{
\item Valid for \code{"ucdpprioconflict"}, \code{"dyadic"}, \code{"battledeaths"}, and \code{"nonstate"} resources.
\item Not valid for the \code{"onesided"} resource.
\item Must be a numeric vector representing conflict IDs.
}}

\item{year}{A vector of years in YYYY format.
\itemize{
\item Valid for all resources.
\item Must be a numeric vector or character vector in the format \code{"YYYY"}.
}}

\item{conflict_incompatibility}{A vector of numeric conflict incompatibility codes.
\itemize{
\item Valid for \code{"ucdpprioconflict"}, \code{"dyadic"}, and \code{"battledeaths"} resources.
\item Not valid for \code{"nonstate"} and \code{"onesided"} resources.
\item Must be a numeric vector representing the type of incompatibility.
}}

\item{dyad}{A vector of numeric dyad IDs.
\itemize{
\item Valid for \code{"gedevents"}, \code{"dyadic"}, \code{"battledeaths"}, and \code{"onesided"} resources.
\item Not valid for \code{"ucdpprioconflict"} and \code{"nonstate"} resources.
\item Must be a numeric vector representing dyad IDs.
}}

\item{org}{A vector of numeric organization IDs.
\itemize{
\item Valid only for the \code{"nonstate"} resource.
\item Not valid for \code{"ucdpprioconflict"}, \code{"dyadic"}, \code{"battledeaths"}, \code{"onesided"}, and \code{"gedevents"} resources.
\item Must be a numeric vector representing organization IDs.
}}

\item{conflict_type}{A vector of numeric conflict type codes.
\itemize{
\item Valid for \code{"ucdpprioconflict"}, \code{"dyadic"}, and \code{"battledeaths"} resources.
\item Not valid for \code{"nonstate"}, \code{"onesided"}, and \code{"gedevents"} resources.
\item Must be a numeric vector representing conflict type codes.
}}

\item{geography}{A string specifying the geographic bounding box in the format "y0 x0,y1 x1".
\itemize{
\item Valid only for the \code{"gedevents"} resource.
\item Format: Two coordinates describing the bounding box (latitude and longitude).
\item Example: \code{"40 -75,41 -73"} for New York City.
}}

\item{start_date}{A string specifying the start date in "YYYY-MM-DD" format.
\itemize{
\item Valid only for the \code{"gedevents"} resource.
\item Must specify the date from which events have definitely happened.
}}

\item{end_date}{A string specifying the end date in "YYYY-MM-DD" format.
\itemize{
\item Valid only for the \code{"gedevents"} resource.
\item Must specify the date up to which events have definitely happened.
}}

\item{type_of_violence}{A vector of numeric violence type codes.
\itemize{
\item Valid only for the \code{"gedevents"} resource.
\item Must be a numeric vector representing types of violence.
}}

\item{actor}{A vector of numeric actor IDs.
\itemize{
\item Valid only for the \code{"gedevents"} resource.
\item Must be a numeric vector representing actor IDs.
}}
}
\value{
A data frame containing the requested data.
}
\description{
This function retrieves a dataset from the Uppsala Conflict Data Program (UCDP) API.
It takes the resource type and version as input parameters, constructs the API request,
and returns the combined results of all pages. The function also validates and filters
the input parameters based on the specified resource to ensure compatibility with the API.
}
\seealso{
\url{https://ucdp.uu.se/apidocs/}
}
